+++ internal/testsuite/otaa_test.go
@@ -184,7 +184,10 @@ func TestOTAAScenarios(t *testing.T) {
 							ResultCode: backend.Success,
 						},
 						NwkSKey: &backend.KeyEnvelope{
-							AESKey: lorawan.AES128Key{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1},
+							AESKey: []byte{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1},
+						},
+						AppSKey: &backend.KeyEnvelope{
+							AESKey: []byte{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16},
 						},
 					},
 					DeviceQueueItems: []storage.DeviceQueueItem{
@@ -222,15 +225,18 @@ func TestOTAAScenarios(t *testing.T) {
 					},
 					ExpectedPHYPayload: jaPHY,
 					ExpectedDeviceSession: storage.DeviceSession{
-						MACVersion:            "1.0.2",
-						RoutingProfileID:      rp.ID,
-						DeviceProfileID:       dp.ID,
-						ServiceProfileID:      sp.ID,
-						JoinEUI:               lorawan.EUI64{1, 2, 3, 4, 5, 6, 7, 8},
-						DevEUI:                lorawan.EUI64{2, 2, 3, 4, 5, 6, 7, 8},
-						FNwkSIntKey:           lorawan.AES128Key{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1},
-						SNwkSIntKey:           lorawan.AES128Key{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1},
-						NwkSEncKey:            lorawan.AES128Key{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1},
+						MACVersion:       "1.0.2",
+						RoutingProfileID: rp.ID,
+						DeviceProfileID:  dp.ID,
+						ServiceProfileID: sp.ID,
+						JoinEUI:          lorawan.EUI64{1, 2, 3, 4, 5, 6, 7, 8},
+						DevEUI:           lorawan.EUI64{2, 2, 3, 4, 5, 6, 7, 8},
+						FNwkSIntKey:      lorawan.AES128Key{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1},
+						SNwkSIntKey:      lorawan.AES128Key{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1},
+						NwkSEncKey:       lorawan.AES128Key{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1},
+						AppSKeyEvelope: &storage.KeyEnvelope{
+							AESKey: []byte{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16},
+						},
 						RXWindow:              storage.RX1,
 						EnabledUplinkChannels: []int{0, 1, 2},
 						ExtraUplinkChannels:   map[int]band.Channel{},
@@ -254,13 +260,16 @@ func TestOTAAScenarios(t *testing.T) {
 							ResultCode: backend.Success,
 						},
 						SNwkSIntKey: &backend.KeyEnvelope{
-							AESKey: lorawan.AES128Key{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1},
+							AESKey: []byte{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1},
 						},
 						FNwkSIntKey: &backend.KeyEnvelope{
-							AESKey: lorawan.AES128Key{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 2},
+							AESKey: []byte{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 2},
 						},
 						NwkSEncKey: &backend.KeyEnvelope{
-							AESKey: lorawan.AES128Key{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 3},
+							AESKey: []byte{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 3},
+						},
+						AppSKey: &backend.KeyEnvelope{
+							AESKey: []byte{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16},
 						},
 					},
 					DeviceQueueItems: []storage.DeviceQueueItem{
@@ -299,15 +308,115 @@ func TestOTAAScenarios(t *testing.T) {
 					},
 					ExpectedPHYPayload: jaPHY,
 					ExpectedDeviceSession: storage.DeviceSession{
-						MACVersion:            "1.1.0",
-						RoutingProfileID:      rp.ID,
-						DeviceProfileID:       dp.ID,
-						ServiceProfileID:      sp.ID,
-						JoinEUI:               lorawan.EUI64{1, 2, 3, 4, 5, 6, 7, 8},
-						DevEUI:                lorawan.EUI64{2, 2, 3, 4, 5, 6, 7, 8},
-						SNwkSIntKey:           lorawan.AES128Key{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1},
-						FNwkSIntKey:           lorawan.AES128Key{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 2},
-						NwkSEncKey:            lorawan.AES128Key{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 3},
+						MACVersion:       "1.1.0",
+						RoutingProfileID: rp.ID,
+						DeviceProfileID:  dp.ID,
+						ServiceProfileID: sp.ID,
+						JoinEUI:          lorawan.EUI64{1, 2, 3, 4, 5, 6, 7, 8},
+						DevEUI:           lorawan.EUI64{2, 2, 3, 4, 5, 6, 7, 8},
+						SNwkSIntKey:      lorawan.AES128Key{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1},
+						FNwkSIntKey:      lorawan.AES128Key{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 2},
+						NwkSEncKey:       lorawan.AES128Key{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 3},
+						AppSKeyEvelope: &storage.KeyEnvelope{
+							AESKey: []byte{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16},
+						},
+						RXWindow:              storage.RX1,
+						EnabledUplinkChannels: []int{0, 1, 2},
+						ExtraUplinkChannels:   map[int]band.Channel{},
+						UplinkGatewayHistory:  map[lorawan.EUI64]storage.UplinkGatewayHistory{},
+						RX2Frequency:          config.C.NetworkServer.Band.Band.GetDefaults().RX2Frequency,
+						NbTrans:               1,
+					},
+				},
+				{
+					BeforeFunc: func(tc *otaaTestCase) error {
+						config.C.JoinServer.KEK.Set = []struct {
+							Label string
+							KEK   lorawan.AES128Key `mapstructure:"kek"`
+						}{
+							{
+								Label: "010203",
+							},
+						}
+
+						dp.MACVersion = "1.1.0"
+						return storage.UpdateDeviceProfile(db, &dp)
+					},
+					Name:       "join-request accepted (SNwkSIntKey, FNwkSIntKey, NwkSEncKey with KEK)",
+					RXInfo:     rxInfo,
+					PHYPayload: jrPayload,
+					AppKey:     appKey,
+					JoinServerJoinAnsPayload: backend.JoinAnsPayload{
+						PHYPayload: backend.HEXBytes(jaBytes),
+						Result: backend.Result{
+							ResultCode: backend.Success,
+						},
+						SNwkSIntKey: &backend.KeyEnvelope{
+							KEKLabel: "010203",
+							AESKey:   []byte{246, 176, 184, 31, 61, 48, 41, 18, 85, 145, 192, 176, 184, 141, 118, 201, 59, 72, 172, 164, 4, 22, 133, 211},
+						},
+						FNwkSIntKey: &backend.KeyEnvelope{
+							KEKLabel: "010203",
+							AESKey:   []byte{87, 85, 230, 195, 36, 30, 231, 230, 100, 111, 15, 254, 135, 120, 122, 0, 44, 249, 228, 176, 131, 73, 143, 0},
+						},
+						NwkSEncKey: &backend.KeyEnvelope{
+							KEKLabel: "010203",
+							AESKey:   []byte{78, 225, 236, 219, 189, 151, 82, 239, 109, 226, 140, 65, 233, 189, 174, 37, 39, 206, 241, 242, 2, 127, 157, 247},
+						},
+						AppSKey: &backend.KeyEnvelope{
+							KEKLabel: "lora-app-server",
+							AESKey:   []byte{248, 215, 201, 250, 55, 176, 209, 198, 53, 78, 109, 184, 225, 157, 157, 122, 180, 229, 199, 88, 30, 159, 30, 32},
+						},
+					},
+					DeviceQueueItems: []storage.DeviceQueueItem{
+						{
+							DevEUI:     lorawan.EUI64{2, 2, 3, 4, 5, 6, 7, 8},
+							FRMPayload: []byte{1, 2, 3, 4},
+							FCnt:       10,
+							FPort:      1,
+						},
+					},
+
+					ExpectedJoinReqPayload: backend.JoinReqPayload{
+						BasePayload: backend.BasePayload{
+							ProtocolVersion: backend.ProtocolVersion1_0,
+							SenderID:        "030201",
+							ReceiverID:      "0102030405060708",
+							MessageType:     backend.JoinReq,
+						},
+						MACVersion: "1.1.0",
+						PHYPayload: backend.HEXBytes(jrBytes),
+						DevEUI:     d.DevEUI,
+						DLSettings: lorawan.DLSettings{
+							OptNeg:      true,
+							RX2DataRate: uint8(config.C.NetworkServer.NetworkSettings.RX2DR),
+							RX1DROffset: uint8(config.C.NetworkServer.NetworkSettings.RX1DROffset),
+						},
+						RxDelay: config.C.NetworkServer.NetworkSettings.RX1Delay,
+					},
+					ExpectedTXInfo: gw.TXInfo{
+						MAC:       rxInfo.MAC,
+						Timestamp: &timestamp,
+						Frequency: rxInfo.Frequency,
+						Power:     14,
+						DataRate:  rxInfo.DataRate,
+						CodeRate:  rxInfo.CodeRate,
+					},
+					ExpectedPHYPayload: jaPHY,
+					ExpectedDeviceSession: storage.DeviceSession{
+						MACVersion:       "1.1.0",
+						RoutingProfileID: rp.ID,
+						DeviceProfileID:  dp.ID,
+						ServiceProfileID: sp.ID,
+						JoinEUI:          lorawan.EUI64{1, 2, 3, 4, 5, 6, 7, 8},
+						DevEUI:           lorawan.EUI64{2, 2, 3, 4, 5, 6, 7, 8},
+						SNwkSIntKey:      lorawan.AES128Key{88, 148, 152, 153, 48, 146, 207, 219, 95, 210, 224, 42, 199, 81, 11, 241},
+						FNwkSIntKey:      lorawan.AES128Key{83, 127, 138, 174, 137, 108, 121, 224, 21, 209, 2, 208, 98, 134, 53, 78},
+						NwkSEncKey:       lorawan.AES128Key{152, 152, 40, 60, 79, 102, 235, 108, 111, 213, 22, 88, 130, 4, 108, 64},
+						AppSKeyEvelope: &storage.KeyEnvelope{
+							KEKLabel: "lora-app-server",
+							AESKey:   []byte{248, 215, 201, 250, 55, 176, 209, 198, 53, 78, 109, 184, 225, 157, 157, 122, 180, 229, 199, 88, 30, 159, 30, 32},
+						},
 						RXWindow:              storage.RX1,
 						EnabledUplinkChannels: []int{0, 1, 2},
 						ExtraUplinkChannels:   map[int]band.Channel{},
@@ -331,7 +440,7 @@ func TestOTAAScenarios(t *testing.T) {
 							ResultCode: backend.Success,
 						},
 						NwkSKey: &backend.KeyEnvelope{
-							AESKey: lorawan.AES128Key{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1},
+							AESKey: []byte{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1},
 						},
 					},
 
@@ -410,7 +519,7 @@ func TestOTAAScenarios(t *testing.T) {
 							ResultCode: backend.Success,
 						},
 						NwkSKey: &backend.KeyEnvelope{
-							AESKey: lorawan.AES128Key{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1},
+							AESKey: []byte{16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1},
 						},
 					},
 					DeviceQueueItems: []storage.DeviceQueueItem{